refactor

+++ src/main/java/faang/school/postservice/redis/cache/service/post/PostCacheServiceImpl.java
@@ -1,5 +1,6 @@
 package faang.school.postservice.redis.cache.service.post;
 
+import faang.school.postservice.dto.user.UserDto;
 import faang.school.postservice.redis.cache.entity.PostCache;
 import faang.school.postservice.redis.cache.repository.PostCacheRepository;
 import faang.school.postservice.redis.cache.service.RedisOperations;
@@ -10,7 +11,6 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 @Slf4j
@@ -25,28 +25,38 @@ public class PostCacheServiceImpl implements PostCacheService {
     private final AuthorCacheService authorCacheService;
 
     @Override
-    public CompletableFuture<PostCache> save(PostCache entity, List<Long> subscriberIds) {
+    public void save(PostCache entity) {
 
         entity = redisOperations.updateOrSave(postCacheRepository, entity, entity.getId());
 
-        PostCache finalEntity = entity;
-        authorCacheService.save(entity.getAuthor());
-        subscriberIds.forEach(subscriberId -> feedCacheService.addPostToFeed(finalEntity, subscriberId));
-
         log.info("Saved post with id {} to cache: {}", entity.getId(), entity);
 
-        return CompletableFuture.completedFuture(entity);
+        PostCache finalEntity = entity;
+        CompletableFuture<UserDto> author = authorCacheService.save(entity.getAuthor());
+
+        author.thenAccept(userDto -> userDto.getSubscriberIds()
+                .forEach(subscriberId -> feedCacheService.addPostToFeed(finalEntity, subscriberId))
+        );
     }
 
     @Override
-    public void deleteById(long postId, List<Long> subscriberIds) {
+    public void deleteById(long postId) {
 
         PostCache post = redisOperations.findById(postCacheRepository, postId).orElse(null);
-        redisOperations.deleteById(postCacheRepository, postId);
-
-        subscriberIds.forEach(subscriberId -> feedCacheService.deletePostFromFeed(post, subscriberId));
 
         log.info("Deleted post with id={} from cache", postId);
+
+        if (post != null) {
+
+            redisOperations.deleteById(postCacheRepository, postId);
+
+            if (post.getAuthor() != null) {
+                CompletableFuture<UserDto> author = authorCacheService.getUserDtoByCache(post.getAuthor());
+                author.thenAccept(userDto -> userDto.getSubscriberIds()
+                        .forEach(subscriberId -> feedCacheService.deletePostFromFeed(post, subscriberId))
+                );
+            }
+        }
     }
 
     @Override
